Code cleanup
- remove commented code
- fix compilation warning

+++ src/pass/adaptation/conv_dilation_pass.cpp
@@ -135,10 +135,10 @@ std::array<unsigned short, 4> calcNewPadding(mv::Data::OpListIterator opIt, size
         unsigned int out_width = ceil(newWidth / oldStride[1]);
         auto pad_along_height 	= std::max(int((out_height - 1) * oldStride[0] + kernelShape[mv::KERNEL_HEIGHT] - newHeight), 0);
         auto pad_along_width 	= std::max(int((out_width - 1) * oldStride[1] + kernelShape[mv::KERNEL_WIDTH] - newWidth), 0);
-        unsigned int pad_top 	= pad_along_height / 2;
-        unsigned int pad_bottom 	= pad_along_height - pad_top;
-        unsigned int pad_left 	= pad_along_width / 2;
-        unsigned int pad_right 	= pad_along_width - pad_left;
+        short unsigned int pad_top 	= pad_along_height / 2;
+        short unsigned int pad_bottom 	= pad_along_height - pad_top;
+        short unsigned int pad_left 	= pad_along_width / 2;
+        short unsigned int pad_right 	= pad_along_width - pad_left;
 
         std::array<unsigned short, 4> padding = {pad_left, pad_right, pad_top, pad_bottom};
         return padding;
@@ -152,7 +152,6 @@ void convDilationUsingStorageElementFcn(const mv::pass::PassEntry& pass, mv::Com
     using namespace mv;
 
     mv::OpModel om(model);
-//    mv::DataModel dm(model);
     auto convOps = om.getOps("Conv");
 
     for (auto& opIt : convOps)
@@ -160,22 +159,10 @@ void convDilationUsingStorageElementFcn(const mv::pass::PassEntry& pass, mv::Com
         auto dilationFactor = opIt->get<unsigned>("dilationFactor");
         if (dilationFactor > 1)
         {
-//            auto params = model.getGlobalConfigParams();
-
-//            std::size_t cmxSize = params->get<unsigned>("cmx");
-//            auto nextOp = findSinkLayers(dm, opIt->getOutputTensor(0))[0];
             auto nonDilatedKernel = opIt->getInputTensor(1);
             auto nonDilatedKernelShape = nonDilatedKernel->getShape();
             auto inputTensor = opIt->getInputTensor(0);
-//            auto outputTensor = opIt->getOutputTensor(0);
-//            auto inputShape = inputTensor->getShape();
-//            auto inputTensorMemory = inputShape.totalSize() * std::ceil(inputTensor->getDType().getSizeInBits()/8.0);
-//            bool sliceWith3DDMA = false;
-//            //NOTE: temporary hack, normally the placement of the ops should be the same and the
-//            //attribute sliceWith3DDma/sparsity(disabled) should be set after the g.o, but on that
-//            //case g.o. will compute memory with sparsity in and produce the analogous strategies
-//            if (inputTensorMemory > cmxSize)
-//                sliceWith3DDMA = true;
+
             auto name = opIt->getName();
 
             auto originalShape = inputTensor->getShape();
@@ -185,8 +172,6 @@ void convDilationUsingStorageElementFcn(const mv::pass::PassEntry& pass, mv::Com
             size_t height = originalShape[mv::IO_HEIGHT_DIMENSION];
             size_t sliceWidth = std::ceil(((double)width)/dilationFactor);
             size_t sliceHeight = std::ceil(((double)height)/dilationFactor);
-            //size_t sliceWidth = originalShape[mv::IO_WIDTH_DIMENSION]/dilationFactor;
-            //size_t sliceHeight = originalShape[mv::IO_HEIGHT_DIMENSION]/dilationFactor;
             std::array<unsigned short, 4> padding = calcNewPadding(opIt, sliceWidth, sliceHeight);
 
             //Create sub dilated convs
@@ -224,7 +209,6 @@ void convDilationUsingStorageElementFcn(const mv::pass::PassEntry& pass, mv::Com
                 inputSlots.push_back(sinkFlow->get<std::size_t>("sinkInput"));
             }
 
-//            auto dtype = opIt->get<mv::DType>("dType");
             auto quantParams = opIt->get<mv::QuantizationParams>("quantParams");
             auto opId = opIt->get<unsigned>("opId");
 
@@ -233,9 +217,6 @@ void convDilationUsingStorageElementFcn(const mv::pass::PassEntry& pass, mv::Com
             mv::Data::TensorIterator concatIt;
             std::vector<mv::Data::TensorIterator> subConvsPerColumn;
             std::vector<mv::Data::TensorIterator> firstLevelConcats;
-//            bool needSparse2SparseOp = false;
-//            if (outputTensorMemory > 0) //Hack always true for now until we fix the implicit join case
-//            {
             pass.log(mv::Logger::MessageType::Debug, "Dilated Conv concat of concat case " +  name);
 
             for (size_t i = 0; i < dilationFactor; i++)
@@ -262,92 +243,6 @@ void convDilationUsingStorageElementFcn(const mv::pass::PassEntry& pass, mv::Com
                 opsToLink[j]->setInputTensor(concatIt, inputSlots[j], false);
                 om.defineFlow(concatIt, opsToLink[j], inputSlots[j]);
             }
-//            }
-//            else
-//            {
-//                pass.log(mv::Logger::MessageType::Debug, "Dilated Conv Implicit Join case " +  name);
-
-//                // Specify that next layer requires sparse input
-//                // At least for now until we have a way to convert a tensor with storage elements into a dense one
-//                // Assuming that this will be done after SSD-512
-//                if (nextOp->isSparsityConsumer())
-//                    nextOp->set<bool>("forcedToHaveActivationSparsityDueToDilatedConv", true);
-//                else
-//                    needSparse2SparseOp = true;
-//                concatIt = om.implicitJoin(subConvs,
-//                    "HW",
-//                    dtype,
-//                    quantParams,
-//                    name + "dilatedjoin");
-//                if (!needSparse2SparseOp)
-//                    for (unsigned j = 0; j < opsToLink.size(); ++j)
-//                    {
-//                        opsToLink[j]->setInputTensor(concatIt, inputSlots[j], false);
-//                        om.defineFlow(concatIt, opsToLink[j], inputSlots[j]);
-//                    }
-//            }
-//            om.getSourceOp(concatIt)->set<unsigned>("opId", opId);
-
-
-//            //NOTE: for now i will place a neutral z-major convolution just for re-order
-//            //but under chat with runtime we can make it work without computations, with by-passing
-//            if (needSparse2SparseOp)
-//            {
-//                pass.log(mv::Logger::MessageType::Debug, "Dilated Conv needSparse2SparseOp is on " +  name);
-
-//                mv::Shape weightsShape({1, 1, outputShape[mv::IO_CHANNEL_DIMENSION], outputShape[mv::IO_CHANNEL_DIMENSION]});
-//                std::vector<int64_t> weightsData(weightsShape.totalSize());
-//                for (unsigned k = 0; k < weightsShape[mv::KERNEL_OUTPUT_CHANNELS]; ++k)
-//                {
-//                    for (unsigned c = 0; c < weightsShape[mv::KERNEL_INPUT_CHANNELS]; ++c)
-//                    {
-//                        for (unsigned h = 0; h < weightsShape[mv::KERNEL_HEIGHT]; ++h)
-//                        {
-//                            for (unsigned w = 0; w < weightsShape[mv::KERNEL_WIDTH]; ++w)
-//                            {
-//                                const size_t idx = (k * weightsShape[mv::KERNEL_INPUT_CHANNELS] * weightsShape[mv::KERNEL_WIDTH] * weightsShape[mv::KERNEL_HEIGHT]) +
-//                                                    (c * weightsShape[mv::KERNEL_WIDTH] * weightsShape[mv::KERNEL_HEIGHT]) +
-//                                                    (h * weightsShape[mv::KERNEL_WIDTH]) +
-//                                                    w;
-//                                if (c == k)
-//                                    weightsData[idx] = 255;
-//                                else
-//                                    weightsData[idx] = 0;
-//                            }
-//                        }
-//                    }
-//                }
-
-//                auto sparse2SparseWeights = om.constantInt(
-//                    weightsData,
-//                    weightsShape,
-//                    nonDilatedKernel->get<mv::DType>("dType"),
-//                    mv::Order("NHWC"),
-//                    {{ZP_RANGE_0_1},{SCALE_RANGE_0_1},{0.0},{1.0}},
-//                    inputTensor->getName() + "_DilatedSparse2SparseWeights");
-
-//                auto sparse2SparseConv = om.conv(concatIt,
-//                        sparse2SparseWeights,
-//                        {1, 1},
-//                        {0, 0, 0, 0},
-//                        1,
-//                        1,
-//                        nonDilatedKernel->get<mv::DType>("dType"),
-//                        concatIt->get<mv::QuantizationParams>("quantParams"),
-//                        inputTensor->getName() + "_DilatedSparse2SparseConv");
-
-//                auto sparse2SparseConvOp = om.getSourceOp(sparse2SparseConv);
-//                auto sparse2SparseWeightsOp = om.getSourceOp(sparse2SparseWeights);
-//                sparse2SparseConvOp->set<unsigned>("opId", om.getSourceOp(concatIt)->get<unsigned>("opId"));
-//                sparse2SparseWeightsOp->set<unsigned>("opId", om.getSourceOp(concatIt)->get<unsigned>("opId"));
-
-//                for (unsigned j = 0; j < opsToLink.size(); ++j)
-//                {
-//                    opsToLink[j]->setInputTensor(sparse2SparseConv, inputSlots[j], false);
-//                    om.defineFlow(sparse2SparseConv, opsToLink[j], inputSlots[j]);
-//                }
-//                sparse2SparseConvOp->set<bool>("forcedToHaveActivationSparsityDueToDilatedConv", true);
-//            }
         }
     }
